Tried to fix 32-bit builds.
* css/StylePropertySet.cpp:
(WebCore::StylePropertySet::asText):



Canonical link: https://commits.webkit.org/99651@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@112214 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/css/StylePropertySet.cpp
@@ -716,12 +716,12 @@ String StylePropertySet::asText() const
                 continue;
             if (!shorthandPropertyAppeared.get(shortPropertyIndex))
                 value = getPropertyValue(shorthandPropertyID);
-            shorthandPropertyAppeared.ensureSizeAndSet(shortPropertyIndex, numCSSProperties);
+            shorthandPropertyAppeared.ensureSizeAndSet(static_cast<size_t>(shortPropertyIndex), static_cast<size_t>(numCSSProperties));
         }
 
         if (!value.isNull()) {
             propertyID = shorthandPropertyID;
-            shorthandPropertyUsed.ensureSizeAndSet(shortPropertyIndex, numCSSProperties);
+            shorthandPropertyUsed.ensureSizeAndSet(static_cast<size_t>(shortPropertyIndex), static_cast<size_t>(numCSSProperties));
         } else
             value = prop.value()->cssText();
